Move ROM to rom.go
