+++ ww.go
@@ -3,8 +3,6 @@ package ww
 import (
 	"context"
 	"crypto/rand"
-	_ "embed"
-	"encoding/hex"
 	"errors"
 	"io"
 	"runtime"
@@ -12,40 +10,9 @@ import (
 	"capnproto.org/go/capnp/v3"
 	"github.com/tetratelabs/wazero"
 	"github.com/tetratelabs/wazero/imports/wasi_snapshot_preview1"
-	"lukechampine.com/blake3"
 )
 
 const Version = "0.1.0"
-
-//go:embed internal/rom/main.wasm
-var defaultROM []byte
-
-// ROM is an immutable, read-only memory segment containing WASM
-// bytecode.  It is uniquely identified by its hash.
-type ROM struct {
-	bytecode []byte
-}
-
-func DefaultROM() ROM {
-	return ROM{defaultROM}
-}
-
-func Read(r io.Reader) (rom ROM, err error) {
-	rom.bytecode, err = io.ReadAll(r)
-	return
-}
-
-func (rom ROM) Hash() [64]byte {
-	return blake3.Sum512(rom.bytecode)
-}
-
-// String returns the BLAKE3-512 hash of the ROM, truncated to the
-// first 8 bytes.  It is intended as a human-readable symbol.  Use
-// the Hash() method to verify integrity.
-func (rom ROM) String() string {
-	hash := rom.Hash()
-	return hex.Dump(hash[:8])
-}
 
 // Ww is the execution context for WebAssembly (WASM) bytecode,
 // allowing it to interact with (1) the local host and (2) the
